// web/src/lib/api/generated/models/internalApiHandlersEnrichmentFailureResponse.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * ViewRA Media Server API
 * Self-hosted media server for movies, TV shows, and music
 * OpenAPI spec version: 0.0.1
 */

export interface InternalApiHandlersEnrichmentFailureResponse {
  attempts?: number
  error_category?: string
  error_message?: string
  id?: number
  last_attempt_at?: string
  max_attempts?: number
  media_id?: number
  media_type?: string
  stage?: string
  title?: string
}